requirements.txt
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
Django==2.1.4
django-rest-auth==0.9.3
djangorestframework==3.9.0
httpie==1.0.2
idna==2.8
mysqlclient==1.3.12
Pillow==5.4.0
Pygments==2.3.1
pytz==2018.7
requests==2.21.0
six==1.12.0
South==1.0.2
urllib3==1.24.1
